% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board.R
\name{board}
\alias{board}
\title{Creating board object}
\usage{
board(mat = NULL, n = 5, p = 0.25)
}
\arguments{
\item{mat}{matrix, default to be null, if not null, mat will be used for board}

\item{n}{integer, dimension of the board}

\item{p}{number between 0 and 1, probability of having 0's}
}
\value{
board object
}
\description{
Creating board object
}
\examples{
board(n = 5, p = .25)
}
