% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{change_1}
\alias{change_1}
\title{change a string to 0 or 1}
\usage{
change_1(s)
}
\arguments{
\item{s}{a string}
}
\value{
a vector of 0 or 1
}
\description{
change a string to 0 or 1
}
\examples{
change_1(". * . . *")
}
