% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{change_vec}
\alias{change_vec}
\title{change a vector of strings containing * and .}
\usage{
change_vec(v, d)
}
\arguments{
\item{v}{a vector of strings of * and .}

\item{d}{size of the matrix}
}
\value{
board that contains only 1 and 0
}
\description{
change a vector of strings containing * and .
}
\examples{
change_vec <- function(c("."), 1)
}
