% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{char_row_to_board_row}
\alias{char_row_to_board_row}
\title{Translate A Character Vector to A Board Row}
\usage{
char_row_to_board_row(board_row_as_char)
}
\arguments{
\item{board_row_as_char}{A string vector of length n,
which is the dimension of the board. Only strings that encode recognized board
values are translated into numerics, otherwise NA is returned to denote that
the string does not correspond to any meaningful board value.}
}
\value{
A vector of 0, 1, or NA's of the same length as the input vector.
}
\description{
Translate A Character Vector to A Board Row
}
