% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{char_to_board}
\alias{char_to_board}
\title{Translate a snippet of character strings to a board object}
\usage{
char_to_board(board_as_char)
}
\arguments{
\item{boards_as_chars}{Expected input format:
----
4
* . . *
. . * *
. * . .
. . . .
----}
}
\value{
A board object as encoded by the inputted snippet.
}
\description{
Translate a snippet of character strings to a board object
}
