% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_contain}
\alias{check_contain}
\title{check_contain function; checks whether the vector contains only "*" and "."}
\usage{
check_contain(v)
}
\arguments{
\item{v}{a vector to be check, if TRUE contains only "*" and "."; returns FALSE otherwise}
}
\value{
returns TRUE if the vector contains on;y "*" and "."; returns FALSE otherwise
}
\description{
check_contain function; checks whether the vector contains only "*" and "."
}
