% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percolate.R
\name{check_possible}
\alias{check_possible}
\title{check the whole board if there is more possible water flows}
\usage{
check_possible(x)
}
\arguments{
\item{x}{board obj}
}
\value{
boolean TRUE if there is more possible water flows
}
\description{
check the whole board if there is more possible water flows
}
