% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_to_matrix}
\alias{convert_to_matrix}
\title{conver_to_matrix}
\usage{
convert_to_matrix(txt_mat, n)
}
\arguments{
\item{txt_mat}{matrix in text form}

\item{n}{nrow of matrix}
}
\value{
matrix in numeric form
}
\description{
conver_to_matrix
}
