% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_board}
\alias{create_board}
\title{create_board, create single board from single_board_lines specifically between '----'}
\usage{
create_board(single_board_lines)
}
\arguments{
\item{single_board_lines}{}
}
\value{
single board objects
}
\description{
create_board, create single board from single_board_lines specifically between '----'
}
