% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percolate.R
\name{fill}
\alias{fill}
\title{fill, helper function for percolate board}
\usage{
fill(board, filled)
}
\arguments{
\item{board}{board}

\item{filled}{location that are already filled}
}
\description{
fill, helper function for percolate board
}
