% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percolate.R
\name{flooded}
\alias{flooded}
\title{flooded}
\usage{
flooded(mat, nr, nc)
}
\arguments{
\item{mat}{board(matrix)}

\item{nr}{row index}

\item{nc}{column index}
}
\value{
board(matrix) if cell of given row index, and column index percolates
}
\description{
flooded
}
\examples{
flooded(board(), 1, 1)
}
