% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percolate.R
\name{flow}
\alias{flow}
\title{flow}
\usage{
flow(mat)
}
\arguments{
\item{mat}{takes board(matrix) and start percolate from first row}
}
\value{
mat (class board, matrix) after the flow ends and reflect all the flooded cell
}
\description{
flow
}
\examples{
flow(board())
}
