% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_board_mat}
\alias{generate_board_mat}
\title{generates a matrix of 0 and 1}
\usage{
generate_board_mat(n = 5, p = 0.25)
}
\arguments{
\item{n}{integer, the dimension of the square matrix}

\item{p}{a number between 0, 1, it's the probability to have 0's}
}
\value{
board with n rows and n columns and floor(p*n^2) 0's, the rest are 1's
}
\description{
generates a matrix of 0 and 1
}
\examples{
generate_board_mat(n = 5, p = 0.25)
}
