% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percolate.R
\name{get_filled_location}
\alias{get_filled_location}
\title{helper function for percolate board, get the locations that are filled}
\usage{
get_filled_location(board)
}
\arguments{
\item{board}{a board}
}
\value{
row and column
}
\description{
helper function for percolate board, get the locations that are filled
}
