% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_valid}
\alias{is_valid}
\title{test whether a matrix a valid}
\usage{
is_valid(mat)
}
\arguments{
\item{mat}{matrix}
}
\value{
boolean values. True if matrix is square and only contains 0, 1, 2
}
\description{
test whether a matrix a valid
}
\examples{
is_valid(matrix(0, 1, 2))
}
