% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_valid_board_lines}
\alias{is_valid_board_lines}
\title{is_valid_board_lines, check whether lines between '----' are valid}
\usage{
is_valid_board_lines(n, board_elements)
}
\arguments{
\item{n}{first element of single_board_lines, supposed to be a size n of a board}

\item{board_elements}{characters}
}
\value{
TRUE if valid, FALSE if invalid
}
\description{
is_valid_board_lines, check whether lines between '----' are valid
}
