% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_valid_dry}
\alias{is_valid_dry}
\title{Dry Board Validity Check}
\usage{
is_valid_dry(board_mat)
}
\arguments{
\item{board_mat}{A board that already passes the is_valid checks}
}
\value{
TRUE if the board is valid and dry
}
\description{
Checks if a valid board is dry -- contains only 0 and 1's
}
\examples{
is_valid_dry(board())
}
