% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_valid_lines_to_board}
\alias{is_valid_lines_to_board}
\title{is_valid_lines_to_board, helper function for read_boards}
\usage{
is_valid_lines_to_board(n, mat)
}
\arguments{
\item{n}{row and col}

\item{mat}{matrix}
}
\value{
True if valid
}
\description{
is_valid_lines_to_board, helper function for read_boards
}
