% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lines_to_board}
\alias{lines_to_board}
\title{lines_to_board, helper function for read_boards}
\usage{
lines_to_board(line)
}
\arguments{
\item{line}{turn the lines into boards}
}
\value{
turn the lines into boards
}
\description{
lines_to_board, helper function for read_boards
}
