% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mat_list_to_board}
\alias{mat_list_to_board}
\title{mat_list_to_board() function that converts list with sizes and string matrices to list of boards}
\usage{
mat_list_to_board(mat_list)
}
\arguments{
\item{mat_list}{has two components of equivalent length: 
n, which specifies the size of each board and 
mat, which is a string version of each board (of length n^2, for the corresponding n)}
}
\value{
a list of boards converted from the strings, or NA if incorrectly formatted
}
\description{
mat_list_to_board() function that converts list with sizes and string matrices to list of boards
}
