% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percolate-package.R
\docType{package}
\name{percolate-package}
\alias{percolate}
\alias{percolate-package}
\title{percolate: Percolation Simulation}
\description{
This package simulates the percolation function.
}
\author{
\strong{Maintainer}: Eden Hu \email{edenhu11@gmail.com} (YOUR-ORCID-ID)

}
\keyword{internal}
