% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percolate.R
\name{percolate}
\alias{percolate}
\title{add percolate method to percolate board}
\usage{
percolate(x)
}
\arguments{
\item{x}{board obj}
}
\value{
percolated board
}
\description{
add percolate method to percolate board
}
