% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percolate.R
\name{percolate_fill}
\alias{percolate_fill}
\title{Board Percolation Single Iteration}
\usage{
percolate_fill(curr_board)
}
\arguments{
\item{curr_board}{A board with the latest board state.}
}
\value{
An updated board by performing a single percolation step on each entry
of the inputted board. If no more water could be filled into the board, the
returned board should be identical to the inputted board.
}
\description{
For each entry on the current board, percolate a single step (up, down, left,
right) if the entry contains water. This is a helper function of percolate.board
}
\examples{
curr_result_board <- board()
updated_result_board <- percolate_fill(curr_result_board)
if (sum(updated_result_board != curr_result_board) == 0){
   cat("Percolation completed!")
} else {
   cat("More percolation steps required to fill the board!")
}

}
