% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percolate.R
\name{percolate_side}
\alias{percolate_side}
\title{percolate_side}
\usage{
percolate_side(mat, n, start)
}
\arguments{
\item{mat}{board}

\item{n}{num rows of board}

\item{start}{row number}
}
\value{
percolate all squares around the current square
}
\description{
percolate_side
}
