% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percolate.R
\name{percolate_step}
\alias{percolate_step}
\title{Board Percolation Single Entry}
\usage{
percolate_step(board, i, j)
}
\arguments{
\item{board}{A board with the latest board state.}

\item{i}{Row index of the square to be percolated.}

\item{j}{Column index of the square to be percolated.}
}
\value{
A board with the updated board state. An identical board will be returned
if water could not be filled into the specified location on board.
}
\description{
Fill up the specified location with water, if possible (which means the entry is
not blocked). This is a helper function of percolate_fill
}
\examples{
percolate_step(board(), 1, 1)
}
