% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percolate.R
\name{percolate_top}
\alias{percolate_top}
\title{percolate_top}
\usage{
percolate_top(mat, n, start)
}
\arguments{
\item{mat}{board}

\item{n}{num rows of board}

\item{start}{row number}
}
\value{
percolate the current row from the top down
}
\description{
percolate_top
}
