% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot.board}
\alias{plot.board}
\title{plot the board}
\usage{
\method{plot}{board}(x)
}
\arguments{
\item{x}{board object}
}
\value{
ggplot of the board with squares 0 = black, 1 = white, 2 = blue
}
\description{
plot the board
}
\examples{
plot.board(board(matrix(c(0,1,1,1,0,0,1,1,0,1,0,0,1,0,0,0,0,0,2,2,2,2,2,2,0), 5, 5))
}
