% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board.R
\name{plot_board}
\alias{plot_board}
\title{plot_board}
\usage{
plot_board(x)
}
\arguments{
\item{x}{board}
}
\value{
ggplot graphic of the inputted board (water, empty blocks, and blocked blocks)
}
\description{
plot_board
}
