% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_board}
\alias{read_board}
\title{read_board}
\usage{
read_board(txt, n)
}
\arguments{
\item{txt}{lines of matrix symbols}

\item{n}{nrow of matrix}
}
\value{
calls error if improper formatting, otherwise return conversion of matrix
}
\description{
read_board
}
