% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_boards}
\alias{read_boards}
\title{read in a file of matrices}
\usage{
read_boards(file)
}
\arguments{
\item{file}{link}
}
\value{
list of transformed board objs
}
\description{
read in a file of matrices
}
\examples{
read_boards("https://raw.githubusercontent.com/benjaminleroy/36-350-summer-data/master/Week5/percolation_write_example.txt")
}
