% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scrape_mat}
\alias{scrape_mat}
\title{scrape_mat() takes a string with n and string matrix (separated by new lines) and returns a list containing n and mat}
\usage{
scrape_mat(mat_obj)
}
\arguments{
\item{mat_obj}{in the format of n (a positive integer), followed by the string matrix}
}
\value{
a list containing n and mat
}
\description{
scrape_mat() takes a string with n and string matrix (separated by new lines) and returns a list containing n and mat
}
