% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{string_to_board}
\alias{string_to_board}
\title{string_to_board() a helper function to convert string boards to actual boards if possible, or NA if not}
\usage{
string_to_board(list_mat)
}
\arguments{
\item{list_mat}{a list specifying the size and content of the board to be created}
}
\value{
either a matrix numeric version of the board or NA
}
\description{
string_to_board() a helper function to convert string boards to actual boards if possible, or NA if not
}
