% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transform_one}
\alias{transform_one}
\title{transforms a string of * and . to 0 and 1}
\usage{
transform_one(s)
}
\arguments{
\item{s}{string, containing * and .}
}
\value{
vector of 0 and 1
}
\description{
transforms a string of * and . to 0 and 1
}
\examples{
transform_one("***.")
}
