% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transform_vec}
\alias{transform_vec}
\title{transform a vector of strings containing * and .}
\usage{
transform_vec(v, d)
}
\arguments{
\item{v}{vector, of strings of * and .}

\item{d}{integer, dimension of the matrix}
}
\value{
board obj, of the matrix containing only 0 and 1
}
\description{
transform a vector of strings containing * and .
}
\examples{
transform_vec <- function(c("."), 1)
}
