% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board.R
\name{board}
\alias{board}
\title{An object with superclass "matrix" and subclass "board"}
\usage{
board(mat = NULL, n = 5, p = 0.25)
}
\arguments{
\item{mat}{a matrix (if valid, a square matrix that contains only 0s, 1s, and 2s)}

\item{n}{a positive integer denote the size of the board}

\item{p}{a number between 0 and 1 that denotes the fraction of the n^2 squares are blocked}
}
\value{
'board' sub - class (matrix primary class) object; if matrix input paramter is incorrect, return an error
}
\description{
An object with superclass "matrix" and subclass "board"
}
\examples{
board(generate_board_mat())
}
