% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board.R
\name{board}
\alias{board}
\title{board(): used as class constructor}
\usage{
board(mat = NULL, n = 5, p = 0.25)
}
\arguments{
\item{mat}{optional, requires valid board structure if given}

\item{n}{must be a positive integer, specifies size of the board}

\item{p}{specifies proportion of board blocked, must be between 0 and 1}
}
\value{
valid board object with attributes n and p
}
\description{
board(): used as class constructor
}
\examples{
board(generate_board_mat())
board(n=6, p=0.3)
}
