% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board.R
\name{board}
\alias{board}
\title{Class constructor function for the class "board".}
\usage{
board(mat = NULL, n = 5, p = 0.25)
}
\arguments{
\item{mat}{If not NULL, contains a board matrix.}

\item{n}{Dimension of the square board matrix. Used only when mat == NULL.}

\item{p}{A float between 0 to 1 that denotes the fraction
of the n^2 squares are blocked. Used only when mat == NULL.}
}
\value{
The generated board object per inputted instructions.
}
\description{
Class constructor function for the class "board".
}
\examples{
board(mat=generate_board_mat())
}
