% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board.R
\name{board}
\alias{board}
\title{make a board object}
\usage{
board(mat = NULL, n = 5, p = 0.25)
}
\arguments{
\item{mat}{the board mastrix has default value n=5, p=0.25}

\item{n}{number of trails}

\item{p}{probality}
}
\value{
an object that is borad
}
\description{
make a board object
}
\examples{
board(5, 0.25)
}
