% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board.R
\name{board}
\alias{board}
\title{board}
\usage{
board(mat = NULL, n = 5, p = 0.25)
}
\arguments{
\item{mat}{matrix into a board}

\item{n}{if there is no matrix, n by n matrix}

\item{p}{if there is no matrix, p is the percentage of the squared that are blocked}
}
\value{
a matrix object, inherited from the matrix class, attr n and p
}
\description{
board
}
