% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board.R
\name{board}
\alias{board}
\title{Makes an object board b based on parameters}
\usage{
board(mat = NULL, n = 5, p = 0.25)
}
\arguments{
\item{mat, }{the matrix if povided}

\item{n, }{the dimension of the matrix (row/col)}

\item{p, }{fraction of squares to be shaded}
}
\value{
An object of class board with subclass matrix and attributes n and p
}
\description{
Makes an object board b based on parameters
}
\examples{
board(n = 8, p = 0.75)
board(n = 10, p = 0.8)
}
