% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board.R, R/utils.R
\name{board}
\alias{board}
\title{board}
\usage{
board(mat = NULL, n = 5, p = 0.25)

board(mat = NULL, n = 5, p = 0.25)
}
\arguments{
\item{mat}{board matrix}

\item{n}{number of rows and columns of desired board}

\item{p}{proportion of blocked squares in board}

\item{mat}{board matrix}

\item{n}{number of rows and columns of desired board}

\item{p}{proportion of blocked squares in board}
}
\value{
board object

board object
}
\description{
board

board
}
