% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board.R
\name{board}
\alias{board}
\title{Title board}
\usage{
board(mat = NULL, n = 5, p = 0.25)
}
\arguments{
\item{mat}{takes matrix default NULL}

\item{n}{defines size of matrix when mat is NULL}

\item{p}{a number between 0 and 1 that denotes the fraction of the `n^2` squares are blocked}
}
\value{
board class objects, if given matrix is incorrect, it returns errors
}
\description{
Title board
}
