% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_contain}
\alias{check_contain}
\title{check if a string that only contains * and . and space}
\usage{
check_contain(s)
}
\arguments{
\item{s}{string that contains only * and . and space}
}
\value{
boolean TRUE it only contains * and . and space
}
\description{
check if a string that only contains * and . and space
}
\examples{
check_contain("*.***")
}
