% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percolate.R
\name{check_possible}
\alias{check_possible}
\title{Check if there is more possible open in the baords}
\usage{
check_possible(x)
}
\arguments{
\item{x}{board object}
}
\value{
boolean true or false
}
\description{
Check if there is more possible open in the baords
}
\examples{
check_possible(board(matrix(1, 10, 10)))
}
