% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_board_mat}
\alias{generate_board_mat}
\title{generate_board_mat()}
\usage{
generate_board_mat(n = 5, p = 0.25)
}
\arguments{
\item{n}{specifies the size of one side of the board (must be positive integer)}

\item{p}{specifies the proportion of the board that is blocked (between 0 an 1)}
}
\value{
an nxn board with p-proportion squares blocked (delineated by 0's and 1's)
}
\description{
generate_board_mat()
}
\examples{
generate_board_mat()
generate_board_mat(n = 8, p = 0.75)
}
