% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_board_mat}
\alias{generate_board_mat}
\title{Generate a legal board per inputted instructions.}
\usage{
generate_board_mat(n = 5, p = 0.25)
}
\arguments{
\item{n}{A positive integer denoting the dimension of the square board.}

\item{p}{A float between 0 to 1 that denotes the fraction
of the n^2 squares are blocked}
}
\value{
The n-by-n board with floor(p*n^2) blocked squares
}
\description{
Generate a legal board per inputted instructions.
}
\examples{
generate_board_mat(n = 8, p = 0.75)

}
