% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_board_mat}
\alias{generate_board_mat}
\title{create a board by matrix}
\usage{
generate_board_mat(n = 5, p = 0.25)
}
\arguments{
\item{n}{, the length size of the board}

\item{p}{the ratio of the balck borad}
}
\value{
a new borad
}
\description{
create a board by matrix
}
\examples{
generate_board_mat(5,0.25)
generate_board_mat(n = 8, p = 0.75)
generate_board_mat(n = 8, p = 0)
}
