% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_board_mat}
\alias{generate_board_mat}
\title{Returns a randomly generated matrix for the percolation}
\usage{
generate_board_mat(n = 5, p = 0.25)
}
\arguments{
\item{n, }{Dimension of row/col for matrix}

\item{p, }{Fraction of squares shaded}
}
\value{
n x n matrix that contains floor(p * n^2) shaded squares
}
\description{
Returns a randomly generated matrix for the percolation
}
\examples{
generate_board_mat(n = 8, p = 0.75)
}
