% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_board_mat}
\alias{generate_board_mat}
\title{generate_board_mat()}
\usage{
generate_board_mat(n = 5, p = 0.25)
}
\arguments{
\item{n}{number of rows and columns}

\item{p}{proportion of blocked out squares}
}
\value{
n x n board with the required number of blocked out units
}
\description{
generate_board_mat()
}
