% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_valid}
\alias{is_valid}
\title{is_valid()}
\usage{
is_valid(mat)
}
\arguments{
\item{mat}{specifies our test variables}
}
\value{
whether or not `mat` is a valid board
}
\description{
is_valid()
}
\examples{
is_valid(generate_board_mat())
is_valid(generate_board_mat(n=1))
}
