% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_valid}
\alias{is_valid}
\title{Checks if a board matrix is valid. By definition, a valid board matrix is
a square matrix containing only 0 or 1. If the input is invalid, this function
throws an error, otherwise it returns TRUE.}
\usage{
is_valid(board_mat)
}
\arguments{
\item{board_mat}{Board matrix to be checked.}
}
\value{
Either throws an error or returns TRUE
}
\description{
Checks if a board matrix is valid. By definition, a valid board matrix is
a square matrix containing only 0 or 1. If the input is invalid, this function
throws an error, otherwise it returns TRUE.
}
\examples{
is_valid(generate_board_mat())
}
