% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_valid}
\alias{is_valid}
\title{check if it is a valid board}
\usage{
is_valid(mat)
}
\arguments{
\item{mat, }{a matrix}
}
\value{
true or error
}
\description{
check if it is a valid board
}
\examples{
is_valid(c(matrix(c(1,2,3,4), nrow = 2)))
is_valid(generate_board_mat())
is_valid(generate_board_mat(n=1))
}
