% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_valid}
\alias{is_valid}
\title{Checks if matrix is valid}
\usage{
is_valid(mat)
}
\arguments{
\item{mat, }{a matrix}
}
\value{
Return TRUE if mat is valid, otherwise throw an error
}
\description{
Checks if matrix is valid
}
\examples{
is_valid(generate_board_mat())
is_valid(generate_board_mat(n=1))
}
