% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board.R, R/utils.R
\name{is_valid}
\alias{is_valid}
\title{is_valid}
\usage{
is_valid(mat)

is_valid(mat)
}
\arguments{
\item{mat}{matrix represening the board}

\item{mat}{matrix represening the board}
}
\value{
TRUE if proper board

TRUE if proper board
}
\description{
is_valid

is_valid
}
