% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_valid}
\alias{is_valid}
\title{Title is_valid}
\usage{
is_valid(mat)
}
\arguments{
\item{mat}{matrix}
}
\value{
error if matrix does not qualifies, return TRUE if pass
}
\description{
Title is_valid
}
\examples{
is_valid(generate_board_mat())
}
