% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percolate-package.R
\docType{package}
\name{percolate-package}
\alias{percolate}
\alias{percolate-package}
\title{percolate: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (YOUR-ORCID-ID)

}
\keyword{internal}
