% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percolate-package.R
\docType{package}
\name{percolate-package}
\alias{percolate}
\alias{percolate-package}
\title{percolate: Allows user to create boards and see if board percolates}
\description{
Use can create a board and perform different functions on the board, like percolate, plot etc
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (YOUR-ORCID-ID)

}
\keyword{internal}
