% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percolate.R
\name{percolate.board}
\alias{percolate.board}
\title{percolate.board function; determines result_board given input "board" object and returns the boolean result of percolation}
\usage{
\method{percolate}{board}(x)
}
\arguments{
\item{x}{a "board" object (superclass - matrix, subclass - board)}
}
\value{
a list of the result_board and the result i.e. whether the "board" percolates (TRUE) or not (FALSE)
}
\description{
percolate.board function; determines result_board given input "board" object and returns the boolean result of percolation
}
\examples{
percolate(board_example_list[[1]]), percolate(board_example_list[[2]]), percolate(board_example_list[[3]])
}
