% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percolate.R
\name{percolate.board}
\alias{percolate.board}
\title{percolate.board()}
\usage{
\method{percolate}{board}(start_board)
}
\arguments{
\item{start_board}{specifies the board to test percolation of}
}
\value{
a list containing the board once water has been filled, and whether the board percolates
}
\description{
percolate.board()
}
\examples{
mat_example_list <- list(matrix(c(1,1,1,1,0,
                                  0,0,0,1,0,
                                  1,1,1,1,0,
                                  0,1,0,0,0,
                                  0,1,1,1,1), 5, 5),
                        matrix(c(1,1,1,1,0,
                                 0,0,0,1,0,
                                 0,1,1,1,0,
                                 0,1,0,0,0,
                                 0,1,1,1,1), 5, 5),
                        matrix(c(1,1,1,1,0,
                                 0,0,0,1,0,
                                 0,1,1,0,0,
                                 0,1,0,0,0,
                                 0,1,1,1,1), 5, 5))

board_example_list <- lapply(mat_example_list, FUN=board)
results <- lapply(board_example_list, percolate)

for (result in results){
  print(result$result)
}

b1 <- plot(board_example_list[[1]])
b2 <- plot(board_example_list[[2]])
b3 <- plot(board_example_list[[3]])
b4 <- plot(results[[1]]$`result_board`)
b5 <- plot(results[[2]]$`result_board`)
b6 <- plot(results[[3]]$`result_board`)

grid.arrange(b1, b2, b3, b4, b5, b6, nrow=2)
}
