% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percolate.R
\name{percolate.board}
\alias{percolate.board}
\title{returns the board with water filled}
\usage{
\method{percolate}{board}(x)
}
\arguments{
\item{x}{board obj}
}
\value{
board with all possible open squares filled
}
\description{
returns the board with water filled
}
