% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percolate.R
\name{percolate.board}
\alias{percolate.board}
\title{percolate the boards}
\usage{
\method{percolate}{board}(x)
}
\arguments{
\item{x}{board object}
}
\value{
a percolate matrix
}
\description{
percolate the boards
}
\examples{
percolate(board(matrix(1, 10, 10)))
}
