% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percolate.R
\name{percolate.board}
\alias{percolate.board}
\title{percolate.board}
\usage{
\method{percolate}{board}(board)
}
\arguments{
\item{board}{a board object}
}
\value{
a list, result_board is the percolated board, result is True if last row is filled
}
\description{
percolate.board
}
