% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percolate.R
\name{percolate.board}
\alias{percolate.board}
\title{percolate_board}
\usage{
percolate.board(start_board)
}
\arguments{
\item{start_board}{board}
}
\value{
list: result board after percolation attempt, T if percolation successful, else F
}
\description{
percolate_board
}
