% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percolate.R
\name{percolate.board}
\alias{percolate.board}
\title{Title percolate.board}
\usage{
\method{percolate}{board}(start_board)
}
\arguments{
\item{start_board}{board(matrix) of initial stage}
}
\value{
list of result_board(after the percolate end), and result whether it is percolate or not
}
\description{
Title percolate.board
}
\examples{
mat_example_list <- list(matrix(c(1,1,1,1,0,
0,0,0,1,0,
1,1,1,1,0,
0,1,0,0,0,
0,1,1,1,1), 5, 5),
matrix(c(1,1,1,1,0,
        0,0,0,1,0,
        0,1,1,1,0,
        0,1,0,0,0,
        0,1,1,1,1), 5, 5),
matrix(c(1,1,1,1,0,
        0,0,0,1,0,
        0,1,1,0,0,
        0,1,1,1,1), 5, 5))
## turn matrices to board
example_boards <- lapply(mat_example_list, board)
## apply boards to percolate.board
laply(example_boards, percolate.board)
}
