% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percolate.R
\name{percolate}
\alias{percolate}
\title{UseMethod for percolate; indicates that for a "board" class object, percolate is the method used for percolate.board}
\usage{
percolate(x)
}
\arguments{
\item{x}{a "board" object (superclass - matrix, subclass - board)}
}
\value{
a list of the result_board and the result i.e. whether the "board" percolates (TRUE) or not (FALSE)
}
\description{
UseMethod for percolate; indicates that for a "board" class object, percolate is the method used for percolate.board
}
