% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percolate.R
\name{percolate}
\alias{percolate}
\title{percoalte function}
\usage{
percolate(x, ...)
}
\arguments{
\item{x}{board}

\item{...}{}
}
\value{
a borad
}
\description{
percoalte function
}
\examples{
percolate(myboard())
}
