% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percolate.R
\name{percolate}
\alias{percolate}
\title{percolate}
\usage{
percolate(x, ...)
}
\arguments{
\item{x}{board}
}
\value{
list: result board after percolation attempt, T if percolation successful, else F
}
\description{
percolate
}
