% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percolate.R
\name{percolate}
\alias{percolate}
\title{percolate}
\usage{
percolate(mat, n)
}
\arguments{
\item{mat}{board}

\item{n}{num rows of the board}
}
\value{
does a percolation iteration of the board
}
\description{
percolate
}
