% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percolate.R
\name{percolate}
\alias{percolate}
\title{percolate}
\usage{
percolate(start_board)
}
\arguments{
\item{start_board}{}
}
\value{
percolate.board result (since there is no percolate for matrix,
        board inherited from matrix should call percolate instead)
}
\description{
percolate
}
\examples{
percolate(board())
}
