% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot.board}
\alias{plot.board}
\title{plot.board method; plots a "board" class object}
\usage{
\method{plot}{board}(x, grid = TRUE)

\method{plot}{board}(x, grid = TRUE)
}
\arguments{
\item{x}{a "board" object (superclass - matrix, subclass - board)}

\item{grid}{a boolean, to decide whether the plot should contain dashed grid lines or not}

\item{x}{a "board" object (superclass - matrix, subclass - board)}
}
\value{
a tile plot containing squares that represent the elements of "board" (superclass matrix); black for 0s, white for 1s, and light blue for 2s

a tile plot containing squares that represent the elements of "board" (superclass matrix); black for 0s, white for 1s, and light blue for 2s; if grid = TRUE, adds dashed gray lines to geom_tile
}
\description{
plot.board method; plots a "board" class object

plot.board method updated; includes a grid which if TRUE is represneted visually by gray dashed lines
}
\examples{
plot(board_example), plot(board_example2), plot(board_example3), plot(board_example4)
plot(board_example, grid = TRUE), plot(board_example2, grid = TRUE), plot(board_example3, grid = TRUE), plot(board_example4, grid = TRUE)
}
