% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot.board}
\alias{plot.board}
\title{plot.board(), plotting function for valid boards}
\usage{
\method{plot}{board}(x)
}
\arguments{
\item{x}{specifies the board to be plotted; must be a valid board object}
}
\value{
the plot of the board x
}
\description{
plot.board(), plotting function for valid boards
}
\examples{
board_example <- board(matrix(c(0,1,1,1,0,
0,1,1,0,1,
0,0,1,0,0,
0,0,0,2,2,
2,2,2,2,0), 5, 5))
plot.board(board_example)
}
