% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot.board}
\alias{plot.board}
\title{Board Visualization}
\usage{
\method{plot}{board}(board, grid)
}
\arguments{
\item{board:}{a board object. An error will occur if the inputted board is
invalid.}

\item{grid:}{if TRUE, the squares are plotted with grey, dashed outlines}
}
\value{
Returns the ggplot object visualizing the board.
}
\description{
Plot a board based on the inputted board matrix.
Different values in the matrix represent squares of different colors.
0: black; 1: white; 2: lightblue
}
\examples{
board_example <- board(matrix(c(0,1,1,1,0,
                                0,1,1,0,1,
                                0,0,1,0,0,
                                0,0,0,2,2,
                                2,2,2,2,0), 5, 5))
note the matrix will be the transpose of what you see here (due to byrow = F)
plot(board_example, grid=TRUE)

board_example2 <- board(matrix(c(0,1,1,1,0,
                                 0,1,1,0,1,
                                 0,0,1,0,0,
                                 0,0,0,1,1,
                                 1,1,1,1,0), 5, 5))
plot(board_example2, grid=TRUE)

board_example3 <- board(matrix(c(0,2,2,2,0,
                                 0,2,2,0,2,
                                 0,0,2,0,0,
                                 0,0,0,2,2,
                                 2,2,2,2,0), 5, 5))
plot(board_example3, grid=TRUE)

board_example4 <- board(    # board of size 10 (correctly percolated)
    matrix(c(2, 0, 0, 2, 0, 0, 2, 2, 0, 0,
             2, 2, 2, 0, 0, 2, 2, 0, 0, 1,
             0, 2, 0, 0, 0, 0, 2, 0, 0, 0,
             0, 2, 2, 2, 0, 0, 0, 0, 0, 0,
             0, 2, 0, 0, 0, 0, 1, 0, 0, 0,
             2, 2, 2, 0, 1, 0, 0, 1, 1, 1,
             0, 2, 2, 0, 1, 1, 0, 0, 1, 0,
             1, 0, 2, 0, 0, 0, 0, 1, 1, 0,
             1, 0, 0, 1, 1, 0, 0, 0, 0, 0,
             0, 1, 1, 0, 0, 1, 0, 1, 0, 0),
             10, 10))
plot(board_example4, grid=TRUE)
}
