% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot.board}
\alias{plot.board}
\title{make a functon to plot board}
\usage{
\method{plot}{board}(x)
}
\arguments{
\item{x}{a board}
}
\value{
to pot a graph
}
\description{
make a functon to plot board
}
\examples{
plot.board(matrix)
plot(board(generate_board_mat()))
}
