% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot.board}
\alias{plot.board}
\title{plot.board}
\usage{
\method{plot}{board}(board, grid = TRUE)
}
\arguments{
\item{board}{board object}

\item{grid}{True, dashed lines indicate the grid}
}
\value{
a plot of the board
}
\description{
plot.board
}
