% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot.board}
\alias{plot.board}
\title{Plots the board x}
\usage{
\method{plot}{board}(x)
}
\arguments{
\item{x, }{the board}
}
\value{
A plot of the board, with the correct colors and dimensions
}
\description{
Plots the board x
}
