% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot.board}
\alias{plot.board}
\title{plot.board}
\usage{
\method{plot}{board}(x, grid = FALSE)
}
\arguments{
\item{x}{matrix}
}
\value{
visualize the matrix with 0 = black, 1 = white, 2 = ligtblue
}
\description{
plot.board
}
\examples{
run plot(board())
}
