% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_boards}
\alias{read_boards}
\title{read_boards() function; takes in a filepath to a text file and outputs a list of matrices that the text file represents}
\usage{
read_boards(file)
}
\arguments{
\item{file}{text file that contains boards;}
}
\value{
list of matrices that the text file represents
}
\description{
read_boards() function; takes in a filepath to a text file and outputs a list of matrices that the text file represents
}
\examples{
read_boards("https://raw.githubusercontent.com/benjaminleroy/36-350-summer-data/master/Week5/percolation_write_example.txt")
}
