% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_boards}
\alias{read_boards}
\title{Translate Text File to Board Objects}
\usage{
read_boards(file)
}
\arguments{
\item{file}{Path to text files that will be read in using readLines(). The vector
of character strings will contain correctly encoded boards, with a separator
"----" placed before and after each board.

If the file's overal format does not meet the expectation, an error will be thrown.
If the file is formatted correctly overall, but a board has invalid encoding,
no error will occur but the board will be returned as NA rather than a board object.}
}
\value{
A list of board objects (or NA for invalid boards).
}
\description{
Translate Text File to Board Objects
}
\examples{
read_boards("https://raw.githubusercontent.com/benjaminleroy/36-350-summer-data/master/Week5/percolation_write_example.txt")
}
