% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_boards}
\alias{read_boards}
\title{Read a file and translate back to baord}
\usage{
read_boards(x)
}
\arguments{
\item{x}{file link}
}
\value{
list of transformed board objs
}
\description{
Read a file and translate back to baord
}
\examples{
read_boards("https://raw.githubusercontent.com/benjaminleroy/36-350-summer-data/master/Week5/percolation_write_example.txt")
}
