% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_boards}
\alias{read_boards}
\title{read_boards}
\usage{
read_boards(file)
}
\arguments{
\item{file}{denote the text file that contains our boards}
}
\value{
board class objects, if lines do not meet these specifications, return NA
}
\description{
read_boards
}
\examples{
read_boards("https://raw.githubusercontent.com/benjaminleroy/36-350-summer-data/master/Week5/percolation_write_test.txt")
}
