% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transform_one}
\alias{transform_one}
\title{transform_one function; transforms a string of "*" to 0 and "." to 1}
\usage{
transform_one(s)
}
\arguments{
\item{s}{a string, containing "*" and "."}
}
\value{
vector of 0 and 1
}
\description{
transform_one function; transforms a string of "*" to 0 and "." to 1
}
\examples{
transform_one("***.")
}
