% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transform_vec}
\alias{transform_vec}
\title{transform_vec function; transforms a vector of strings containing "*" and "." to a "board" object}
\usage{
transform_vec(v, d)
}
\arguments{
\item{v}{vector of strings of "*" and "."}

\item{d}{integer, dimensions of matrix}
}
\value{
"board" object, superclass "matrix" containing only 0s and 1s
}
\description{
transform_vec function; transforms a vector of strings containing "*" and "." to a "board" object
}
\examples{
transform_vec(c("."), 1), transform_vec(c("*"), 1)
}
